examples/12_music_processing/pyproject.toml
[project]
name = "12-music-processing"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.56.0",
    "numpy>=2.3.5",
    "pydub>=0.25.1",
    "python-dotenv>=1.2.1",
    "rice-agents",
    "ricedb>=0.1.0",
    "scipy>=1.16.3",
]

[tool.uv.sources]
rice-agents = { workspace = true, editable = true }
